+++ lib/src/actions/effects.dart
@@ -11,6 +11,17 @@ class EffectModifySquare extends ActionEffect {
   const EffectModifySquare(this.square, this.content);
 }
 
+/// Sets the custom state variable at index [variable] to [value].
+/// You can set a number of variables equal to the size of your board, so on an
+/// 8x8 board, the highest [variable] is 63.
+/// [value] can be up to 16383, or 2^46 if you assume your code will never
+/// execute on a 32-bit vm.
+class EffectSetCustomState extends ActionEffect {
+  final int variable;
+  final int value;
+  const EffectSetCustomState(this.variable, this.value);
+}
+
 /// Causes [piece] to be added to [player]'s hand.
 class EffectAddToHand extends ActionEffect {
   final int player;